Added dumps/loads cache
